requirements.txt
altair==5.5.0
aiohttp==3.13.1
aiodns==3.5.0
aiohappyeyeballs==2.6.1
annotated-types==0.7.0
anthropic==0.71.0
anyio==4.11.0
async==0.6.2
asyncpg==0.30.0
attrs==23.2.0
Authlib==1.6.5
backoff==2.2.1
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cozepy==0.19.0
cryptography==46.0.3
distro==1.9.0
docstring_parser==0.17.0
dydantic==0.0.8
filelock==3.20.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.41.1
googleapis-common-protos==1.70.0
grpcio==1.75.1
h11==0.16.0
h2==4.3.0
hf-xet==1.1.10
hpack==4.1.0
httpcore==1.0.9
httpx==0.27.2
huggingface-hub==0.35.3
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==0.3.27
langchain-core==0.3.79
langchain-openai==0.3.35
langchain-text-splitters==0.3.11
langfuse==3.7.0
langgraph==0.6.11
langgraph-checkpoint==2.1.2
langgraph-checkpoint-postgres==2.0.25
langgraph-prebuilt==0.6.5
langgraph-sdk==0.2.9
langgraph_reducer==0.0.2
langmem==0.0.29
langsmith==0.4.37
MarkupSafe==3.0.3
mpmath==1.3.0
narwhals==2.8.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.4
openai==2.5.0
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.3
ormsgpack==1.11.0
packaging==25.0
pandas==2.3.3
pdfminer.six==20250506
pillow==11.3.0
portalocker==3.2.0
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg-pool==3.2.6
psycopg2-binary==2.9.11
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.2
pydantic_core==2.41.4
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
qdrant-client==1.15.1
referencing==0.37.0
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.1
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.44
streamlit==1.50.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
toml==0.10.2
torch==2.9.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.57.1
trustcall==0.0.39
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==14.2
wheel==0.45.1
wrapt==1.17.3
xxhash==3.6.0
PyPDF2==3.0.1
arxiv==2.2.0
feedparser==6.0.12
sgmllib3k==1.0.0
spacy==3.8.7
spacypdfreader==0.3.2
pdf2image==1.17.0
pytesseract==0.3.13
xx_sent_ud_sm @ https://github.com/explosion/spacy-models/releases/download/xx_sent_ud_sm-3.8.0/xx_sent_ud_sm-3.8.0-py3-none-any.whl#sha256=1b5fb9321723b395f76458d2a13b7a3673c537eb58c0a9df3778f4e0950180b4